# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GuestRsvpStatus = typeof GuestRsvpStatus[keyof typeof GuestRsvpStatus];


export const GuestRsvpStatus = {
  going: 'going',
  maybe: 'maybe',
  not_going: 'not_going',
  pending: 'pending',
} as const;

export interface Guest {
  id: number;
  name: string;
  /** @nullable */
  googleId?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  rsvpStatus: GuestRsvpStatus;
  createdAt: string;
}

export type GuestInputRsvpStatus = typeof GuestInputRsvpStatus[keyof typeof GuestInputRsvpStatus];


export const GuestInputRsvpStatus = {
  going: 'going',
  maybe: 'maybe',
  not_going: 'not_going',
  pending: 'pending',
} as const;

export interface GuestInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  googleId?: string;
  avatarUrl?: string;
  rsvpStatus?: GuestInputRsvpStatus;
}

export type RsvpUpdateRsvpStatus = typeof RsvpUpdateRsvpStatus[keyof typeof RsvpUpdateRsvpStatus];


export const RsvpUpdateRsvpStatus = {
  going: 'going',
  maybe: 'maybe',
  not_going: 'not_going',
} as const;

export interface RsvpUpdate {
  rsvpStatus: RsvpUpdateRsvpStatus;
}

export interface RsvpSummary {
  going: number;
  maybe: number;
  notGoing: number;
  pending: number;
  total: number;
}

export interface Post {
  id: number;
  guestId: number;
  guestName: string;
  /** @nullable */
  guestAvatarUrl?: string | null;
  content: string;
  replyCount: number;
  createdAt: string;
}

export interface PostInput {
  guestId: number;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

export interface Reply {
  id: number;
  postId: number;
  guestId: number;
  guestName: string;
  /** @nullable */
  guestAvatarUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface ReplyInput {
  guestId: number;
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
}