# lib/api-zod/src/generated/types/guest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */
import type { GuestRsvpStatus } from './guestRsvpStatus';

export interface Guest {
  id: number;
  name: string;
  /** @nullable */
  googleId?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  rsvpStatus: GuestRsvpStatus;
  createdAt: Date;
}