# lib/api-zod/src/generated/types/guestInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */
import type { GuestInputRsvpStatus } from './guestInputRsvpStatus';

export interface GuestInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
  googleId?: string;
  avatarUrl?: string;
  rsvpStatus?: GuestInputRsvpStatus;
}