# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  guestId: number;
  guestName: string;
  /** @nullable */
  guestAvatarUrl?: string | null;
  content: string;
  replyCount: number;
  createdAt: Date;
}