# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  guestId: number;
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}