# lib/api-zod/src/generated/types/reply.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */

export interface Reply {
  id: number;
  postId: number;
  guestId: number;
  guestName: string;
  /** @nullable */
  guestAvatarUrl?: string | null;
  content: string;
  createdAt: Date;
}