# lib/api-zod/src/generated/types/rsvpSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */

export interface RsvpSummary {
  going: number;
  maybe: number;
  notGoing: number;
  pending: number;
  total: number;
}