# lib/api-zod/src/generated/types/rsvpUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Renker Niver de Bolso — Birthday party guest confirmation API
 * OpenAPI spec version: 0.1.0
 */
import type { RsvpUpdateRsvpStatus } from './rsvpUpdateRsvpStatus';

export interface RsvpUpdate {
  rsvpStatus: RsvpUpdateRsvpStatus;
}